[css] Add -webkit-tap-highlight-color:transparent
Chrome on mobile displays a blue highlight for any clicked element,
which is visually distracting for things like the clickable divs on the
home page and unnecessary, as all links on the site have visual feedback
of some kind upon click.

No change in behaviour for Firefox on Android, which doesn't show a
highlight on tap by default.