[js][css] Rotate footer image once on hover but restart animation upon click
Previously, the footer image would rotate continuously when hovering.
This worked fine when using a mouse, but on mobile the footer image is
stuck rotating after clicking once. This change moves it so the image
rotates just once, but the animation will restart if it finished when
clicking. This gives behaviour that works well across all devices.
Unfortunately it the JS code to restart a CSS animation is a bit of a
hack (removing the relative ID/class and forcing a reflow).