[js] Add both mouseover and touchstart listeners
I can't seem to find a clear indication of whether there are cases when
one might be called and not the other (e.g. a system with both touch and
mouse inputs). As the handlers can tolerate being called twice, just
register both and let that happen.

MDN recommends using preventDefault in a touch handler to prevent mouse
events, but this can't be called from a passive event handler.