[gen][doc] Add mechanism for indicating error in local build and document systemd-user based rebuild
Using the systemd unit file in the README, it's easy to have darkhttpd
serve a version of the site, and have it regenerated upon file changes.
If there is an error in the generation process, a file will be created
in the serving directory that is checked by a snippet of JS. If the
_had_error file is present then the background is changed to a pinkish
red as a visual indicator.